feat: move search and query params from mixtapes page to mixtapes store to keep it
