+++ pages/mixtapes/index.vue
@@ -1,29 +1,14 @@
 <script lang="ts" setup>
 import { useDisplay } from "vuetify/lib/framework.mjs"
-import { MixtapeExt, MixtapesTags, Tag } from "~~/types/supatypes"
+import { MixtapeExt, Tag } from "~~/types/supatypes"
 
 definePageMeta({ middleware: ["auth"] })
 
-interface TagFilter {
-  id: number
-  exclude?: boolean
-}
-
-interface Filters {
-  tags: TagFilter[]
-}
-
 const { actAs } = useProfileStore()
-const { fetchMixtapes, deleteMixtape } = useMixtapesStore()
-const { data: mixtapes } = useMixtapesStoreRefs()
+const { fetchMixtapes, deleteMixtape, filters } = useMixtapesStore()
+const { data: mixtapes, search, itemsPerPage, page } = useMixtapesStoreRefs()
 const { fetchTags } = useTagsStore()
 const { data: tags } = useTagsStoreRefs()
-const search = ref(null)
-const filters: Filters = reactive({
-  tags: [],
-})
-const itemsPerPage = ref(20)
-const page = ref(1)
 const idToDelete = ref<string | number | null>(null)
 const openConfirm = ref(false)
 const { smAndUp, mdAndUp, update } = useDisplay()
